+++ extensions/bitwarden/src/types/passwords.ts
@@ -1,6 +1,10 @@
+import { asOptionKeys } from "~/utils/types";
+
 export type PasswordType = "password" | "passphrase";
 
-export interface PasswordOptions {
+export type PasswordOptions = {
+  /** Do not generate a passphrase */
+  passphrase?: false;
   /** Include uppercase characters */
   lowercase?: boolean;
   /** Include lowercase characters */
@@ -15,9 +19,23 @@ export interface PasswordOptions {
   minNumber?: string;
   /** Minimum number of special characters */
   minSpecial?: string;
-}
+};
 
-export interface PassphraseOptions {
+/** Keys for the password options to be used at runtime */
+export const PASSWORD_OPTION_KEYS: readonly string[] = asOptionKeys<PasswordOptions>()([
+  "passphrase",
+  "lowercase",
+  "uppercase",
+  "number",
+  "special",
+  "length",
+  "minNumber",
+  "minSpecial",
+]);
+
+export type PassphraseOptions = {
+  /** Generate a passphrase */
+  passphrase?: true;
   /** Number of words */
   words?: string;
   /** Word separator */
@@ -26,9 +44,19 @@ export interface PassphraseOptions {
   capitalize?: boolean;
   /** Passphrase includes number */
   includeNumber?: boolean;
-}
+};
+
+/** Keys for the passphrase options to be used at runtime */
+export const PASSPHRASE_OPTION_KEYS: readonly string[] = asOptionKeys<PassphraseOptions>()([
+  "passphrase",
+  "words",
+  "separator",
+  "capitalize",
+  "includeNumber",
+]);
+
 export type PasswordGeneratorOptions = {
   /** Generate a passphrase */
-  passphrase?: boolean;
-} & PasswordOptions &
-  PassphraseOptions;
+  passphrase: boolean;
+} & Omit<PasswordOptions, "passphrase"> &
+  Omit<PassphraseOptions, "passphrase">;