+++ extensions/bitwarden/src/utils/passwords.ts
@@ -2,11 +2,13 @@ import { LocalStorage } from "@raycast/api";
 import { pbkdf2 } from "crypto";
 import { LOCAL_STORAGE_KEY } from "~/constants/general";
 import { DEFAULT_PASSWORD_OPTIONS, REPROMPT_HASH_SALT } from "~/constants/passwords";
-import { PasswordGeneratorOptions } from "~/types/passwords";
+import { PASSPHRASE_OPTION_KEYS, PASSWORD_OPTION_KEYS, PasswordGeneratorOptions } from "~/types/passwords";
 
 export function getPasswordGeneratingArgs(options: PasswordGeneratorOptions): string[] {
+  const validOptions: readonly string[] = options.passphrase ? PASSPHRASE_OPTION_KEYS : PASSWORD_OPTION_KEYS;
   return Object.entries(options).flatMap(([arg, value]) => {
-    if (value == null || value === "") return [];
+    if (!validOptions.includes(arg)) return [];
+    if (value == null || value === "" || value === false) return [];
     if (value === true) return [`--${arg}`];
     return [`--${arg}`, value];
   });